collect/requirements.txt
# Core dependencies for dual scoring collector
requests==2.32.4
opencv-python-headless==4.12.0.88
numpy==2.2.6
pillow==11.3.0
google-generativeai==0.3.2

# Database dependencies
psycopg2-binary==2.9.9
google-cloud-storage==2.10.0

# Additional packages (may not be needed for basic collector)
matplotlib==3.10.3
folium==0.20.0
pandas==2.3.1
pytz==2025.2
transformers==4.54.1
# PyTorch moved to shared base stage
scikit-learn==1.7.1
joblib==1.5.1